+++ familyhelper-finance-sdk/src/main/java/com/dwarfeng/familyhelper/finance/sdk/util/ServiceExceptionCodes.java
@@ -46,7 +46,17 @@ public final class ServiceExceptionCodes {
      * @param exceptionCodeOffset 指定的异常代号的偏移量。
      */
     public static void setExceptionCodeOffset(int exceptionCodeOffset) {
+        // 设置 EXCEPTION_CODE_OFFSET 的值。
         EXCEPTION_CODE_OFFSET = exceptionCodeOffset;
+
+        // 以新的 EXCEPTION_CODE_OFFSET 为基准，更新异常代码的值。
+        ACCOUNT_BOOK_NOT_EXISTS.setCode(offset(0));
+        BANK_CARD_NOT_EXISTS.setCode(offset(10));
+        ILLEGAL_BANK_CARD_STATE.setCode(offset(20));
+        USER_NOT_EXISTS.setCode(offset(30));
+        USER_NOT_PERMITTED.setCode(offset(40));
+        FUND_CHANGE_NOT_EXISTS.setCode(offset(50));
+        ILLEGAL_FUND_CHANGE_STATE.setCode(offset(60));
     }
 
     private ServiceExceptionCodes() {